refactored to use crates / workspace
