+++ crates/generator/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "generator"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde_json = "1.0.127"
+filesystem = { path = "../filesystem" }
+parser = { path = "../parser" }
