+++ crates/generator/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod parameters;
+mod string_formatter;
