+++ crates/generator/src/string_formatter.rs
