+++ crates/parser/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "parser"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde = { version = "1.0.209", features = ["derive"] }
+serde_json = "1.0.127"
+filesystem = { path = "../filesystem" }
+
+[dev-dependencies]
+rstest.workspace = true
