+++ crates/parser/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod schema;
+pub mod parser;
