+++ crates/parser/src/schema/parameter_type.rs
@@ -1,7 +1,6 @@
-use crate::parser::schema::parameter::Parameter;
-use crate::parser::schema::reference::Reference;
 use serde::Deserialize;
-
+use crate::schema::parameter::Parameter;
+use crate::schema::reference::Reference;
 // todo! Order of them matters: https://serde.rs/enum-representations.html#untagged
 
 #[derive(Deserialize, Debug, PartialEq)]
