+++ crates/parser/tests/filesystem_integration_test.rs
@@ -1,5 +1,5 @@
-use razure::filesystem::get_latest_stable_specifications;
-use razure::parser::parse_specification_file;
+use filesystem::filesystem::get_latest_stable_specifications;
+use parser::parser::parse_specification_file;
 
 #[test]
 fn download_azure_gets_latest_specifications() {
