--- src/generator/mod.rs
@@ -1,2 +1,0 @@
-pub mod parameters;
-mod string_formatter;
