--- src/lib.rs
@@ -1,3 +1,0 @@
-pub mod filesystem;
-pub mod generator;
-pub mod parser;
